refactor

+++ services/commerce-gateway/gateway/web/handler/order.go
@@ -76,7 +76,7 @@ func (oh *orderHandler) CreateOrder(c *gin.Context) {
 	// 上記の制約を受けるのはこの関数のみで、実際のorder作成のロジックには影響しない
 	orderLines := []*pb.OrderLine{
 		{
-			Count: int32(req.Count),
+			Count: int64(req.Count),
 			Item: &pb.CatalogItem{
 				Id: req.ItemID,
 			},
